refactor : Kafka 이벤트 기반 알림 유즈케이스를 application 레이어로 분리
